+++ src/hooks/api/useMinerDetailsQuery/index.ts
