+++ src/hooks/api/useMinerDetailsQuery/useMinerDetailsQuery.tsx
