+++ src/hooks/api/useMinerNotificationSettingsMutation/index.ts
