+++ src/hooks/api/useMinerPaymentsQuery/index.ts
