+++ src/hooks/api/useMinerPaymentsQuery/useMinerPaymentsQuery.tsx
