+++ src/hooks/api/useMinerRewardsQuery/index.ts
