+++ src/hooks/api/useMinerRewardsQuery/useMinerRewardsQuery.tsx
