+++ src/hooks/api/useMinerStatsChartQuery/useMinerStatsChartQuery.tsx
