+++ src/hooks/api/useMinerStatsQuery/index.ts
