+++ src/hooks/api/useMinerStatsQuery/useMinerStatsQuery.tsx
