+++ src/pages/MinerDashboard/Header/Stats.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 import { Card, CardGrid, CardTitle } from 'src/components/layout/Card';
 import { useLocalizedActiveCoinValueFormatter } from 'src/hooks/useDisplayReward';
 import useActiveCoinNetworkFee from '@/hooks/useActiveCoinNetworkFee';
-import useMinerStatsQuery from '@/hooks/useMinerStatsQuery';
-import useMinerDetailsQuery from '@/hooks/useMinerDetailsQuery';
+import useMinerStatsQuery from '@/hooks/api/useMinerStatsQuery';
+import useMinerDetailsQuery from '@/hooks/api/useMinerDetailsQuery';
 import useWorkerStatus from '@/hooks/useWorkerStatus';
 import { StatItem } from 'src/components/StatItem';
 import { useLocalStorageState } from 'src/hooks/useLocalStorageState';
