+++ src/rdx/localSettings/localSettings.hooks.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 import { useReduxState } from 'src/rdx/useReduxState';
 import { useDispatch } from 'react-redux';
 import { localSettingsSet } from './localSettings.actions';
-import usePoolCoinsQuery from '@/hooks/usePoolCoinsQuery';
+import usePoolCoinsQuery from '@/hooks/api/usePoolCoinsQuery';
 
 // FIXME: This hook isn't reliable for retrieving the current active coin ticker,
 // since the redux state can be updated many re-renders later
